Refactor st ConsumerGroup is the supervisor
This makes more intuitive sense - one wouldn't really start the manager
process without the supervisor.
